#432 fix delete destination

See merge request de.pronin/p_dp_airline_1!388

+++ airline-project/src/main/java/app/services/DestinationService.java
@@ -3,8 +3,10 @@ package app.services;
 import app.dto.DestinationDto;
 import app.entities.Destination;
 import app.enums.Airport;
+import app.exceptions.DestinationConnectedFlightsException;
 import app.mappers.DestinationMapper;
 import app.repositories.DestinationRepository;
+import app.repositories.FlightRepository;
 import lombok.RequiredArgsConstructor;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.PageRequest;
@@ -19,6 +21,7 @@ import java.util.stream.Collectors;
 public class DestinationService {
 
     private final DestinationRepository destinationRepository;
+    private final FlightRepository flightRepository;
     private final DestinationMapper destinationMapper;
 
     public List<DestinationDto> getAllDestinations() {
@@ -88,6 +91,11 @@ public class DestinationService {
 
     @Transactional
     public void deleteDestinationById(Long id) {
+        if (checkFlightsWithThisDestinationExist(id)) throw new DestinationConnectedFlightsException(id);
         destinationRepository.deleteById(id);
+    }
+
+    public boolean checkFlightsWithThisDestinationExist(Long id) {
+        return !flightRepository.findByDestinationId(id).isEmpty();
     }
 }
